+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Cache for documents loaded via the `$ref` keyword. [#75](https://github.com/Stranger6667/jsonschema-rs/issues/75)
+- Meta schemas for JSON Schema drafts 4, 6, and 7. [#28](https://github.com/Stranger6667/jsonschema-rs/issues/28)
 
 ### Fixed